+++ main.tf
@@ -4,27 +4,18 @@ locals {
   repository_uri = "${data.aws_caller_identity.current.account_id}.dkr.ecr.${var.aws_region}.amazonaws.com/${var.upstream_registry_name}"
 }
 
-# Create a secret in AWS Secrets Manager to store the upstream registry name.
-resource "aws_secretsmanager_secret" "ecr_pullthroughcache" {
+data "aws_secretsmanager_secret" "ecr_pullthroughcache" {
   name = "ecr-pullthroughcache/${var.upstream_registry_name}"
 }
 
-resource "aws_secretsmanager_secret_version" "ecr_pullthroughcache" {
-  secret_id = aws_secretsmanager_secret.ecr_pullthroughcache.id
-  secret_string = jsonencode({
-    username    = "UPDATE-ME"
-    accessToken = "UPDATE-ME"
-  })
-
-  lifecycle {
-    ignore_changes = [secret_string]
-  }
+data "aws_secretsmanager_secret_version" "ecr_pullthroughcache" {
+  secret_id = data.aws_secretsmanager_secret.ecr_pullthroughcache.id
 }
 
 resource "aws_ecr_pull_through_cache_rule" "ecr_pullthroughcache" {
   ecr_repository_prefix = var.upstream_registry_name
   upstream_registry_url = var.upstream_registry_url
-  credential_arn        = aws_secretsmanager_secret.ecr_pullthroughcache.arn
+  credential_arn        = data.aws_secretsmanager_secret.ecr_pullthroughcache.arn
 }
 
 resource "aws_iam_policy" "ecr_pullthroughcache" {
@@ -46,4 +37,25 @@ resource "aws_iam_policy" "ecr_pullthroughcache" {
   })
 
   depends_on = [aws_ecr_pull_through_cache_rule.ecr_pullthroughcache]
+}
+
+resource "kubernetes_secret_v1" "secret" {
+  for_each = toset(var.fallback_namespaces)
+
+  metadata {
+    name      = var.fallback_secret_name
+    namespace = each.key
+  }
+  data = {
+    ".dockerconfigjson" = jsonencode(
+      {
+        "auths" : {
+          "var.upstream_registry_url" : {
+            "auth" : base64encode(data.aws_secretsmanager_secret_version.ecr_pullthroughcache.secret_string)
+          }
+        }
+      }
+    )
+  }
+  type = "kubernetes.io/dockerconfigjson"
 }